push cdk wallet up into the bridge
this code is still kind of broken
but what really bugs me is that it's ugly af

open questions:
is there a better solution to make async cdk calls?
do I want to push the wallet into the bridge?
or would it be better to keep the wallet in the proxy
and only carry the signing key through other structs?