index the blinded secret by the block header hash
This might not be the most efficient implementation
but it's far easier to understand using a feeble human mind

We use the quote field of the cashu spec as an index to the blinded
secrets that are stored in the wallet.
This allows for better separation of concerns by isolating the cashu
state inside the cdk wallet.